+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Test version of a new generic command class. 
+- A new generic command class.
+- Ported price command to new command class. 
 
 ### Changed
 - Improved item guessing with unpublished items in the results.